+++ BackEnd/services/moderation/handle_approved_images.py
@@ -1,4 +1,4 @@
-from utils import completeness_count
+from utils.completeness_counter import completeness_count
 
 
 class ApprovedImagesDeleter:
